requirements.txt
Flask
requests
werkzeug
markdown
matplotlib
numpy
pandas
keras
tensorflow
pillow
flask-caching
python-dotenv
google-generativeai
beautifulsoup4
gunicorn
